+++ futures-test/src/task/noop_spawner.rs
@@ -1,5 +1,4 @@
-use futures_core::future::FutureObj;
-use futures_task::{Spawn, SpawnError};
+use futures_task::{Spawn, SpawnError, FutureObj};
 
 /// An implementation of [`Spawn`](futures_core::task::Spawn) that
 /// discards spawned futures when used.
