+++ futures-test/src/task/record_spawner.rs
@@ -1,5 +1,4 @@
-use futures_core::future::FutureObj;
-use futures_task::{Spawn, SpawnError};
+use futures_task::{Spawn, SpawnError, FutureObj};
 
 /// An implementation of [`Spawn`](futures_core::task::Spawn) that records
 /// any [`Future`](futures_core::future::Future)s spawned on it.
