+++ futures-util/src/task/mod.rs
@@ -18,7 +18,10 @@ pub use self::spawn::{SpawnExt, LocalSpawnExt};
 
 pub use futures_core::task::{Context, Poll, Waker, RawWaker, RawWakerVTable};
 
-pub use futures_task::{Spawn, LocalSpawn, SpawnError};
+pub use futures_task::{
+    Spawn, LocalSpawn, SpawnError,
+    FutureObj, LocalFutureObj, UnsafeFutureObj,
+};
 
 pub use futures_task::noop_waker;
 #[cfg(feature = "std")]
